pyproject.toml
[project]
name = "csesaproj"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "pdfminer-six>=20221105",
    "pypdf2>=3.0.0",
    "python-dotenv>=1.2.1",
]
